requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask-jwt-extended
flask-bcrypt
python-dotenv
